rename component files to use CamelCase
