+++ apps/web/src/lib/layouts/DocsSidebar.svelte
