+++ apps/web/src/routes/(docs)/+layout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import DocsLayout from '$lib/layouts/docs-layout.svelte';
+	import DocsLayout from '$lib/layouts/DocsLayout.svelte';
 	import { navigation } from '$lib/navigation';
 
 	let { children } = $props();
